Replace contextify shim with vm shim for browsers
